feat: Extract product badge/rating logic from product-list into a shared scripts/product-badges.js module
